CACHE_REQ: Move result manipulation into a separate module
This patch is preparing the field for coming up patches where new lookup
logic will be added.

Taking this into consideration let's move the result manipulation code
into a separate module and focus purely in the lookups logic in the main
module.

Related:
https://pagure.io/sssd/sssd/issue/3001
