+++ src/tests/cwrap/Makefile.am
@@ -38,6 +38,7 @@ SSSD_LIBS = \
 
 SSSD_CACHE_REQ_OBJ = \
     ../../../src/responder/common/cache_req/cache_req.c \
+    ../../../src/responder/common/cache_req/cache_req_result.c \
     ../../../src/responder/common/cache_req/cache_req_search.c \
     ../../../src/responder/common/cache_req/cache_req_data.c \
     ../../../src/responder/common/cache_req/plugins/cache_req_common.c \
